requirements.txt
flask
google-ads
flask-cors
pymongo
PyYAML
python-dotenv
requests
